+++ packages/core/src/electron-main/electron-main-application.ts
@@ -507,6 +507,9 @@ export class ElectronMainApplication {
                 backendProcess.on('error', error => {
                     reject(error);
                 });
+                backendProcess.on('exit', () => {
+                    reject(new Error('backend process exited'));
+                })
                 app.on('quit', () => {
                     // Only issue a kill signal if the backend process is running.
                     // eslint-disable-next-line no-null/no-null